2933 : fixed Build: #1 failed tests

+++ Dev/Dev2.Activities.Tests/ActivityTests/FolderReadTests.cs
@@ -183,8 +183,10 @@ namespace Dev2.Tests.Activities.ActivityTests
         [TestCleanup]
         public void DeleteDir()
         {
-            dirHelper.Delete(_inputPath, true);
+            if (!string.IsNullOrEmpty(_inputPath) && dirHelper.Exists(_inputPath))
+            {
+                dirHelper.Delete(_inputPath, true);
+            }
         }
-
     }
 }
